Chapter 13/requirements.txt
# Core LangChain ecosystem
langchain>=0.1.16
langchain-community>=0.0.28
langchain-core>=0.1.41

# Embeddings & LLMs
ollama
#nomic-embed-text
#mxbai-embed-Text

# Vector database
chromadb>=0.4.24

# Re-ranking (cross-encoder support)
sentence-transformers>=2.2.2
scikit-learn>=1.3.0

# Hybrid retriever (BM25)
rank_bm25>=0.2.2

# Text loading & processing
pandas>=2.0.3
tqdm>=4.66.1

# File parsing utilities
python-dotenv>=1.0.1

# Memory and chaining
faiss-cpu>=1.7.4
numpy>=1.26.4

# Optional (nice to have for chunking PDFs, Markdown etc.)
unstructured>=0.10.25
pdfminer.six>=20221105

# CLI and logging
rich>=13.6.0
